python_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch>=2.1.0
diffusers>=0.30.2
transformers>=4.41.0
accelerate>=0.30.0
python-multipart
safetensors
requests
psutil
pillow
python-dotenv